Massive memory optimisations.
Switch from a massive character-by-character tree to using
bloom filters.

For checking if a word exists or not we use a master bloom
filter with two cross-checks[1]. A separate filter is used
to track "roots" (whether a given string can become a valid
word if more characters are added).

For a ~99MB word list this seems to reduce memory usage
from ~2GB to ~100MB.

There is a chance that we will now generate false positives,
but with the cross-checked filters they should be minimal
(and given some of the dubious entries in the word lists
anyway, not very harmful if they do happen.)

Reading in the word list now takes longer, but the resulting
struct can be serialised and reused, which negates this.
(The 99MB word list serialises down to ~19MB, which is also
a nice side effect.)

[1]: http://www.naturalspublishing.com/files/published/53oi45i66w1of2.pdf