Refactor, add compile command, add multiplexing
Refactor to be based around a 'SpellChecker' type as that more
accurately describes what the Bloom-filter-based model is.

Change the main functions (Match/Anagram/FromMorse) to take in
a SpellChecker instead of being a method on it, so they work
the same way as a third-party implementation would.

Add in Multiplex* functions which allow running the same things
in parallel on multiple checkers. The Discord bot uses this to
search a "normal" word list and a backup urban-dictionary word
list, annotating the results.

Add in a command to compile a SpellChecker model from a word list,
and commit a bunch of example models compiled in this way.

As many of these are breaking changes, the package has been
bumped to v2.